requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.7
daphne==4.1.2
Django==5.0.6
django-anymail==10.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-timezone-field==6.1.0
django-user-agents==0.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
Markdown==3.6
msgpack==1.0.8
pexpect==4.9.0
priority==2.0.0
prompt-toolkit==3.0.43
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-c==3.1.19
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tk==0.1.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4
